extend v0 instead of adding v1
