--- docs/json_schemas/fiveg_f1/v1/provider.json
@@ -1,134 +1,0 @@
-{
-  "$defs": {
-    "BaseModel": {
-      "properties": {},
-      "title": "BaseModel",
-      "type": "object"
-    },
-    "FivegF1ProviderAppData": {
-      "properties": {
-        "f1_ip_address": {
-          "description": "IPv4 address of the network interface used for F1 traffic",
-          "examples": [
-            "192.168.70.132"
-          ],
-          "format": "ipvanyaddress",
-          "title": "F1 Ip Address",
-          "type": "string"
-        },
-        "f1_port": {
-          "description": "Number of the port used for F1 traffic",
-          "examples": [
-            2153
-          ],
-          "title": "F1 Port",
-          "type": "integer"
-        },
-        "tac": {
-          "description": "Tracking Area Code",
-          "examples": [
-            1
-          ],
-          "maximum": 16777215,
-          "minimum": 1,
-          "title": "Tac",
-          "type": "integer"
-        },
-        "plmns": {
-          "items": {
-            "$ref": "#/$defs/PLMNConfig"
-          },
-          "title": "Plmns",
-          "type": "array"
-        }
-      },
-      "required": [
-        "f1_ip_address",
-        "f1_port",
-        "tac",
-        "plmns"
-      ],
-      "title": "FivegF1ProviderAppData",
-      "type": "object"
-    },
-    "PLMNConfig": {
-      "properties": {
-        "mcc": {
-          "description": "Mobile Country Code",
-          "examples": [
-            "001",
-            "208",
-            "302"
-          ],
-          "pattern": "[0-9][0-9][0-9]",
-          "title": "Mcc",
-          "type": "string"
-        },
-        "mnc": {
-          "description": "Mobile Network Code",
-          "examples": [
-            "01",
-            "001",
-            "999"
-          ],
-          "pattern": "[0-9][0-9][0-9]?",
-          "title": "Mnc",
-          "type": "string"
-        },
-        "sst": {
-          "description": "Slice/Service Type",
-          "examples": [
-            1,
-            2,
-            3,
-            4
-          ],
-          "maximum": 4,
-          "minimum": 1,
-          "title": "Sst",
-          "type": "integer"
-        },
-        "sd": {
-          "description": "Slice Differentiator",
-          "examples": [
-            1
-          ],
-          "maximum": 16777215,
-          "minimum": 1,
-          "title": "Sd",
-          "type": "integer"
-        }
-      },
-      "required": [
-        "mcc",
-        "mnc",
-        "sst",
-        "sd"
-      ],
-      "title": "PLMNConfig",
-      "type": "object"
-    }
-  },
-  "description": "Provider schema for fiveg_f1.",
-  "properties": {
-    "unit": {
-      "anyOf": [
-        {
-          "$ref": "#/$defs/BaseModel"
-        },
-        {
-          "type": "null"
-        }
-      ],
-      "default": null
-    },
-    "app": {
-      "$ref": "#/$defs/FivegF1ProviderAppData"
-    }
-  },
-  "required": [
-    "app"
-  ],
-  "title": "ProviderSchema",
-  "type": "object"
-}
