--- docs/json_schemas/fiveg_f1/v1/requirer.json
@@ -1,48 +1,0 @@
-{
-  "$defs": {
-    "BaseModel": {
-      "properties": {},
-      "title": "BaseModel",
-      "type": "object"
-    },
-    "FivegF1RequirerAppData": {
-      "properties": {
-        "f1_port": {
-          "description": "Number of the port used for F1 traffic",
-          "examples": [
-            2153
-          ],
-          "title": "F1 Port",
-          "type": "integer"
-        }
-      },
-      "required": [
-        "f1_port"
-      ],
-      "title": "FivegF1RequirerAppData",
-      "type": "object"
-    }
-  },
-  "description": "Requirer schema for fiveg_f1.",
-  "properties": {
-    "unit": {
-      "anyOf": [
-        {
-          "$ref": "#/$defs/BaseModel"
-        },
-        {
-          "type": "null"
-        }
-      ],
-      "default": null
-    },
-    "app": {
-      "$ref": "#/$defs/FivegF1RequirerAppData"
-    }
-  },
-  "required": [
-    "app"
-  ],
-  "title": "RequirerSchema",
-  "type": "object"
-}
