--- interfaces/fiveg_f1/v1/interface.yaml
@@ -1,14 +1,0 @@
-name: fiveg_f1
-internal: true
-
-version: 1
-status: draft
-
-providers:
-  - name: oai-ran-cu-k8s
-    url: https://github.com/canonical/oai-ran-cu-k8s
-requirers:
-  - name: oai-ran-du-k8s
-    url: https://github.com/canonical/oai-ran-du-k8s
-
-maintainer: telco
